RandomNumGenerator Note
This program creates a random number using time, so the chance of the same number appearing twice. The first part of the program just sets 5 integers to "rand()" and prints them out. The second part of the program uses a loop to generate 5 random numbers. I found the loop to give a better range of numbers compared to just setting integers to "rand()"